+++ CHANGELOG.md
@@ -1,11 +1,11 @@
-# Unreleased
+## v1.3.0
 
 ### Features
 
 - Added reflection argument to format callbacks on `rebuildSearch` options.
 - Added new navigation and search functions that play nice when used by multiple plugins. Deprecated old navigation and search functions.
 
-## V1.2.0
+## v1.2.0
 
 ### Features